#762 Branding
TextInputLayout in category chooser
