+++ app/src/main/res/layout/dialog_change_category.xml
@@ -8,6 +8,7 @@
     android:orientation="vertical">
 
     <com.google.android.material.textfield.TextInputLayout
+        android:id="@+id/searchWrapper"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:padding="16dp">
